refactor saleList and saleListItem to use TS
